requirements.txt
flask==2.3.3
llama-cpp-python==0.2.11
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0